Added easter egg (fix #240).
If the user taps on the PixaPencil logo in the 'About' fragment five times within the span of 2 seconds, the `EasterEggFragment` (which contains the `EasterEggView` as its root) will show.

More work has to be done, but I am committing this due to the fact that most of the basic functionality is there for portrait mode.